refactor(ProductPreReceipt.tsx): UI랑 상관 없는 코드 훅으로 분리
